Develop

+++ components/Post/PostDetail/modal/modal.tsx
@@ -8,14 +8,13 @@ const Modal = ({ isOpen, onClose, title, content, tripId }: any) => {
   const [hasJoined, setHasJoined] = useState(false);
   const [isFailed, setIsFailed] = useState(false);
 
-  const getToken = () => {
-    return localStorage.getItem("token");
-  };
-
   /*여행 참가 - post 요청 */
   const onJoin = async () => {
+    if (typeof window === "undefined") {
+      throw new Error("localStorage is not available on the server.");
+    }
     try {
-      const token = getToken();
+      const token = localStorage.getItem("token");
 
       if (token) {
         await api.post(
@@ -44,8 +43,11 @@ const Modal = ({ isOpen, onClose, title, content, tripId }: any) => {
 
   /*여행 취소 - delete 요청 */
   const onCancel = async () => {
+    if (typeof window === "undefined") {
+      throw new Error("localStorage is not available on the server.");
+    }
     try {
-      const token = getToken();
+      const token = localStorage.getItem("token");
 
       if (token) {
         await api.delete(`/api/attend/${tripId}`, {
